moved auth and config folder.
