+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/v3/auth/AuthMethod.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.auth;
+package io.swagger.codegen.v3.auth;
 
 public interface AuthMethod {
     String getType();
